chore: move `pyproject.toml` to repository root
Although more semantically sound, having the `pyproject.toml` file at
`server/pyproject.toml` makes it so that all tools that use it need
extra configuration to find it, which results in extra maintenance on
the side of the development team.

This change moves `pyproject.toml` to the root of the repository,
simplifying the project setup.